requirements.txt
flask==3.0.0
gunicorn==22.0.0
firebase-admin==6.5.0
python-dotenv==1.0.1